review fix: the same pattern applied for Decimal field

+++ djautotask/models.py
@@ -572,7 +572,8 @@ class Phase(TimeStampedModel):
     start_date = models.DateTimeField(blank=True, null=True)
     # due_date is end date in autotask UI
     due_date = models.DateTimeField(blank=True, null=True)
-    estimated_hours = models.FloatField(default=0)
+    estimated_hours = models.DecimalField(default=0.0, decimal_places=2,
+                                          max_digits=6)
     number = models.CharField(blank=True, null=True, max_length=50)
     scheduled = models.BooleanField(default=False)
     last_activity_date = models.DateTimeField(blank=True, null=True)
